requirements.txt
streamlit
torch
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
faiss-cpu
Pillow